moved filterBuilderProps and FilterPredicates into seriesActions
